// src/modules/orders/domain/enums/order_sub_statuses.enum.ts
export enum OrderSubStatuses {
  // Created statuses
  PENDING = 'PENDING',
  IN_PREPARATION = 'IN_PREPARATION',
  READY_TO_SHIP = 'READY_TO_SHIP',

  // In transit statuses
  IN_TRANSIT = 'IN_TRANSIT',
  LAST_MILE_DELIVERY = 'LAST_MILE_DELIVERY',
  FAILED_DELIVERY_ATTEMPT = 'FAILED_DELIVERY_ATTEMPT',
  WAITING_FOR_WITHDRAWAL = 'WAITING_FOR_WITHDRAWAL',

  // Delivered Statuses
  DELIVERED = 'DELIVERED',

  // Not Delivered Statuses - Canceled
  INVALID_ADDRESS = 'INVALID_ADDRESS',
  CARRIER_CANCELED = 'CARRIER_CANCELED',
  BUYER_CANCELLED = 'BUYER_CANCELLED',
  SYSTEM_CANCELLED = 'SYSTEM_CANCELLED',
  SELLER_CANCELLED = 'SELLER_CANCELLED',
}